+++ src/index.ts
@@ -86,7 +86,6 @@ export {
     Spring,
     Keyframes,
     Inertia,
-    Point,
     None,
     EasingFunction,
     TargetAndTransition,
@@ -109,3 +108,4 @@ export {
 export { MotionFeature, FeatureProps } from "./motion/features/types"
 export { GestureHandlers } from "./gestures"
 export { DraggableProps, DragHandlers } from "./behaviours/types"
+export * from "./types/geometry"