+++ CHANGELOG.md
@@ -6,11 +6,16 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
--   Magic Motion
+-   `AnimateSharedLayout`
 
 ### Removed
 
--   `positionTransition` and `layoutTransition` props (use `magic` instead).
+-   `positionTransition` and `layoutTransition` props (use `animate={true}` instead).
+-   `Point` type in favour of `Point2D`.
+
+### Fixed
+
+-   Allow `dragConstraints` to be a ref that is smaller than the draggable element.
 
 ## [1.11.1] 2020-06-16